Give the ability to the computer to choose
Through two functions, the computer can now choose randomly from
rock, paper and scissors. The function getComputerChoice calls on
random() to generate an integer from 0 to 2 and returns a choice.
Ideally an array would work better than switch case, but at this point
in the Odin path I decided to leave it as it is.