Make the game play five rounds in a row
Now the game plays 5 consecutive rounds. If any of the rounds ends in
a draw, it is not counted towards the 5 rounds. This way, the five
rounds will never result in a tie. The game can go for as many rounds
as necessary until the sum of the computer score and human score equal
five. When that is the case, we are sure to have a winner because in case
of draws no points are awarded and the game only stops when all of the
five points are awarded.